requirements.txt
feedparser==6.0.10
python-telegram-bot==20.0